+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All Notable changes to `gent/services-opening-hours` package.
 
+## [0.2.0]
+
+### Added
+
+* DMOH-43: Added method to get a Service by its URI.
+* DMOH-43: Added method to get a Service by its Vesta ID.
+
 ## [0.1.0]
 
 Initial release providing API to get Services & Channels + access the opening
@@ -28,5 +35,6 @@ hours for a Channel.
 * DMOH-17: Added the OpeningHoursPeriod & OpeningHoursPeriodHTML methods to the
   ChannelService.
 
+[0.2.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/StadGent/php_package_services-opening-hours/releases/tag/0.1.0
 [Unreleased]: https://github.com/StadGent/php_package_services-opening-hours/compare/master...develop